typos: fix unexpected behaviour
Enable typos.configPath to be of type path. This way, we can
use excludes from .typos.toml when running
`pre-commit run typos --all-files`. Otherwise, the execution
differs from a normal invocation of typos, which is unexpected
and leads to wrong results.

To not break anything, and to be compliant with the existing
API, I modified configPath to be either a Nix path or a string.

[0]: https://github.com/cachix/pre-commit-hooks.nix/pull/387#issuecomment-1893600631